Rename the sample app package and module name
