+++ sample/src/androidTest/java/bg/devlabs/fullscreenvideoviewsample/ExampleInstrumentedTest.java
@@ -1,4 +1,4 @@
-package bg.devlabs.fullscreenvideoviewexample;
+package bg.devlabs.fullscreenvideoviewsample;
 
 import android.content.Context;
 import android.support.test.InstrumentationRegistry;
