+++ sample/src/main/java/bg/devlabs/fullscreenvideoviewsample/ActionBarActivity.kt
@@ -1,4 +1,4 @@
-package bg.devlabs.fullscreenvideoviewexample
+package bg.devlabs.fullscreenvideoviewsample
 
 import android.content.Context
 import android.content.Intent
