+++ sample/src/main/java/bg/devlabs/fullscreenvideoviewsample/RegularActivity.kt
@@ -1,22 +1,22 @@
-package bg.devlabs.fullscreenvideoviewexample
+package bg.devlabs.fullscreenvideoviewsample
 
+import android.app.Activity
 import android.content.Context
 import android.content.Intent
 import android.os.Bundle
-import android.support.v7.app.AppCompatActivity
-import kotlinx.android.synthetic.main.activity_no_action_bar.*
+import kotlinx.android.synthetic.main.activity_regular.*
 
-class NoActionBarActivity : AppCompatActivity() {
+class RegularActivity : Activity() {
     companion object {
         fun start(context: Context) {
-            val starter = Intent(context, NoActionBarActivity::class.java)
+            val starter = Intent(context, RegularActivity::class.java)
             context.startActivity(starter)
         }
     }
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_no_action_bar)
+        setContentView(R.layout.activity_regular)
         val videoPath = "http://clips.vorwaerts-gmbh.de/VfE_html5.mp4"
         fullscreenVideoView.build(videoPath)
     }
