+++ sample/src/main/res/layout/activity_action_bar.xml
