+++ sample/src/main/res/layout/activity_no_action_bar.xml
