+++ sample/src/main/res/layout/activity_regular.xml
