+++ sample/src/main/res/layout/activity_toolbar.xml
