+++ sample/src/main/res/mipmap-hdpi/ic_launcher_foreground.png
